Refactor translations and scripts
