+++ index.html
@@ -309,6 +309,7 @@
 
   <script src="data.js"></script>
   <script src="storage.js"></script>
+  <script src="translations.js"></script>
   <script src="script.js"></script>
 </body>
 </html>
