+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "lint": "eslint .",
     "check-consistency": "node checkConsistency.js",
+    "normalize": "node normalizeData.js",
     "test": "npm run lint && npm run check-consistency && jest"
   },
   "keywords": [],
