+++ script.js
@@ -1,4 +1,5 @@
 // script.js – Main logic for the Camera Power Planner app
+/* global texts, categoryNames */
 
 const VIDEO_OUTPUT_TYPES = [
   '3G-SDI',
@@ -125,7 +126,8 @@ const storedDevices = loadDeviceData();
 if (storedDevices) {
   devices = storedDevices;
 }
-
+// Normalize various camera properties so downstream logic works with
+// consistent structures and value formats.
 function unifyDevices(data) {
   if (!data || typeof data !== 'object') return;
   const ensureList = (list, defaults) => {
@@ -228,580 +230,17 @@ function unifyDevices(data) {
 
 unifyDevices(devices);
 
-
-// Translation text for English and German
-const texts = {
-  en: {
-    appTitle: "Camera Power Planner",
-    appHeading: "Camera Power Consumption App",
-    tagline: "Plan your camera setup and calculate power consumption & battery life.",
-
-    setupManageHeading: "Manage Setup",
-    deviceSelectionHeading: "Device Selection",
-    resultsHeading: "Results",
-    deviceManagerHeading: "Manage Device Database",
-    batteryComparisonHeading: "Battery Comparison",
-
-    savedSetupsLabel: "Saved Setups:",
-    newSetupOption: "-- New Setup --",
-    setupNameLabel: "Setup Name:",
-    deleteSetupBtn: "Delete",
-    saveSetupBtn: "Save",
-
-    cameraLabel: "Camera:",
-    monitorLabel: "Monitor:",
-    videoLabel: "Wireless Video:",
-    distanceLabel: "Distance Sensor:",
-    batteryLabel: "V-Mount Battery:",
-
-    fizLegend: "FIZ (Follow Focus) Systems",
-    fizMotorsLabel: "FIZ Motors:",
-    fizControllersLabel: "FIZ Controllers:",
-
-    totalPowerLabel: "Total Consumption:",
-    totalCurrent144Label: "Total Current (at 14.4V):",
-    totalCurrent12Label: "Total Current (at 12V):",
-    batteryLifeLabel: "Runtime (estimated):",
-
-    warnPinExceeded: "WARNING: current draw ({current}A at ~12V) exceeds the battery's main output rating ({max}A)!",
-    warnPinNear: "Note: current draw is close to the main output limit ({current}A of {max}A).",
-    warnDTapExceeded: "WARNING: current draw ({current}A) exceeds the battery's D-Tap output ({max}A)!",
-    warnDTapNear: "Note: current draw uses over 80% of the D-Tap capacity ({current}A of {max}A).",
-
-    category_cameras: "Cameras",
-    category_monitors: "Monitors",
-    category_video: "Wireless Video",
-    category_fiz_motors: "FIZ Motors",
-    category_fiz_controllers: "FIZ Controllers",
-    category_fiz_distance: "FIZ Distance",
-    category_batteries: "V-Mount Batteries",
-
-    addDeviceHeading: "Add New Device",
-    categoryLabel: "Category:",
-    deviceNameLabel: "Name:",
-    consumptionLabel: "Consumption (W):",
-    capacityLabel: "Capacity (Wh):",
-    pinLabel: "Max Pin A:",
-    dtapLabel: "Max D-Tap A:",
-    cameraWattLabel: "Power Draw (W):",
-    cameraVoltageLabel: "Voltage Range:",
-    cameraPortTypeLabel: "Port Type:",
-    cameraPlatesLabel: "Battery Plates:",
-    cameraMediaLabel: "Recording Media:",
-    cameraLensMountLabel: "Lens Mount:",
-    powerInputsHeading: "Power Inputs",
-    powerDistributionHeading: "Power Distribution",
-    videoOutputsHeading: "Video Outputs",
-    fizConnectorHeading: "FIZ Connector",
-    mediaHeading: "Media",
-    viewfinderHeading: "Viewfinder",
-    lensMountHeading: "Lens Mount",
-    timecodeHeading: "Timecode",
-    powerDistributionLabel: "Outputs:",
-    videoOutputsLabel: "Outputs:",
-    fizConnectorLabel: "Connectors:",
-    viewfinderLabel: "Viewfinders:",
-    timecodeLabel: "Timecode:",
-    addDeviceBtn: "Add",
-    updateDeviceBtn: "Update", // New key for update button
-    editBtn: "Edit", // New key for Edit button in list
-    deleteDeviceBtn: "Delete", // New key for Delete button in list
-    cancelEditBtn: "Cancel",
-    exportDataBtn: "Export Database",
-    importDataBtn: "Import Database",
-    alertImportSuccess: "Database imported successfully! {num_devices} devices loaded.",
-    alertImportError: "Failed to import database. Invalid file or data format.",
-    confirmExportAndRevert: "Are you sure you want to export the current database AND then revert to the default database? This will overwrite your current saved data with the original defaults.",
-    alertExportAndRevertSuccess: "Database exported and reverted to defaults successfully.",
-
-    placeholder_deviceName: "Device name",
-    placeholder_watt: "e.g. 12.5",
-    placeholder_capacity: "e.g. 98",
-    placeholder_pin: "e.g. 10",
-    placeholder_dtap: "e.g. 5",
-    placeholder_voltage: "11V-34V DC",
-    placeholder_port: "Bat LEMO 8-pin",
-    placeholder_plates: '[{"type":"V-Mount","mount":"native"}]',
-    placeholder_media: '[{"type":"CFast 2.0"}]',
-    placeholder_lensmount: "ARRI PL",
-    placeholder_powerdist: "[{}]",
-    placeholder_videooutputs: "[{}]",
-    placeholder_fizconnector: "[{}]",
-    placeholder_timecode: "[{}]",
-
-    toggleDeviceManager: "Edit Device Data…",
-    hideDeviceManager: "Hide Device Data",
-
-    batteryTableLabel: "Battery",
-    runtimeLabel: "Estimated Runtime (h)",
-    batteryLifeUnit: "hrs",
-
-    noBatterySupports: "No battery can supply this load.",
-
-    alertSetupName: "Please enter a name for the setup.",
-    alertSetupSaved: 'Setup "{name}" saved.',
-    alertNoSetupSelected: "Please select a saved setup to delete.",
-    confirmDeleteSetup: 'Really delete setup "{name}"?',
-    // alertSetupDeleted not used
-    confirmDeleteDevice: 'Really delete device "{name}"?',
-    alertDeviceExists: "A device with this name already exists in this category.",
-    alertDeviceAdded: 'Device "{name}" added to category "{category}".',
-    alertDeviceUpdated: 'Device "{name}" updated in category "{category}".',
-    alertDeviceFields: "Please provide valid values for capacity, Pin A and D-Tap A.",
-    alertDeviceWatt: "Please enter a valid watt value.",
-    alertDeviceName: "Device name cannot be empty.",
-    alertInvalidCameraJSON: "Invalid JSON for camera details",
-
-    // NEW TEXTS FOR SETUP MANAGEMENT START HERE
-    setupActionsHeading: "Setup Actions",
-    exportSetupsBtn: "Export All Setups",
-    importSetupsBtn: "Import Setups",
-    generateOverviewBtn: "Generate Overview",
-    alertNoSetupsToExport: "There are no saved setups to export.",
-    alertImportSetupsSuccess: "Successfully imported {num_setups} setups.",
-    alertImportSetupsError: "Error: Could not import setups. The file may be invalid or corrupted.",
-    alertSelectSetupForOverview: "Please select a saved setup to generate an overview.",
-    overviewTitle: "Setup Overview",
-    exportAndRevertBtn: "Export and Revert to default Database",
-    // NEW TEXTS FOR SETUP MANAGEMENT END HERE
-  },
-  es: {
-    appTitle: "Planificador de Consumo de Cámara",
-    appHeading: "Aplicación de Consumo de Energía de Cámara",
-    tagline: "Planifica tu equipo y calcula consumo y autonomía de la batería.",
-
-    setupManageHeading: "Gestionar Configuración",
-    deviceSelectionHeading: "Selección de Dispositivos",
-    resultsHeading: "Resultados",
-    deviceManagerHeading: "Gestionar Base de Datos",
-    batteryComparisonHeading: "Comparación de Baterías",
-
-    savedSetupsLabel: "Configuraciones Guardadas:",
-    newSetupOption: "-- Nueva Configuración --",
-    setupNameLabel: "Nombre de Configuración:",
-    deleteSetupBtn: "Eliminar",
-    saveSetupBtn: "Guardar",
-
-    cameraLabel: "Cámara:",
-    monitorLabel: "Monitor:",
-    videoLabel: "Video Inalámbrico:",
-    distanceLabel: "Sensor de Distancia:",
-    batteryLabel: "Batería V-Mount:",
-
-    fizLegend: "Sistemas FIZ (Follow Focus)",
-    fizMotorsLabel: "Motores FIZ:",
-    fizControllersLabel: "Controles FIZ:",
-
-    totalPowerLabel: "Consumo Total:",
-    totalCurrent144Label: "Corriente Total (a 14.4V):",
-    totalCurrent12Label: "Corriente Total (a 12V):",
-    batteryLifeLabel: "Autonomía (estimada):",
-
-    warnPinExceeded: "ADVERTENCIA: la corriente ({current}A a ~12V) supera el límite principal ({max}A)!",
-    warnPinNear: "Nota: la corriente está cerca del límite ({current}A de {max}A).",
-    warnDTapExceeded: "ADVERTENCIA: la corriente ({current}A) supera la salida D-Tap ({max}A)!",
-    warnDTapNear: "Nota: la corriente usa más del 80% de la capacidad D-Tap ({current}A de {max}A).",
-
-    category_cameras: "Cámaras",
-    category_monitors: "Monitores",
-    category_video: "Video Inalámbrico",
-    category_fiz_motors: "Motores FIZ",
-    category_fiz_controllers: "Controles FIZ",
-    category_fiz_distance: "Distancia FIZ",
-    category_batteries: "Baterías V-Mount",
-
-    addDeviceHeading: "Añadir Nuevo Dispositivo",
-    categoryLabel: "Categoría:",
-    deviceNameLabel: "Nombre:",
-    consumptionLabel: "Consumo (W):",
-    capacityLabel: "Capacidad (Wh):",
-    pinLabel: "Máx Pin A:",
-    dtapLabel: "Máx D-Tap A:",
-    cameraWattLabel: "Consumo (W):",
-    cameraVoltageLabel: "Rango de Voltaje:",
-    cameraPortTypeLabel: "Tipo de Puerto:",
-    cameraPlatesLabel: "Placas de Batería:",
-    cameraMediaLabel: "Medios de Grabación:",
-    cameraLensMountLabel: "Montura de Lente:",
-    powerInputsHeading: "Entradas de Energía",
-    powerDistributionHeading: "Distribución de Energía",
-    videoOutputsHeading: "Salidas de Video",
-    fizConnectorHeading: "Conector FIZ",
-    mediaHeading: "Medios",
-    viewfinderHeading: "Visor",
-    lensMountHeading: "Montura",
-    timecodeHeading: "Timecode",
-    powerDistributionLabel: "Salidas:",
-    videoOutputsLabel: "Salidas:",
-    fizConnectorLabel: "Conectores:",
-    viewfinderLabel: "Visores:",
-    timecodeLabel: "Timecode:",
-    addDeviceBtn: "Añadir",
-    updateDeviceBtn: "Actualizar",
-    editBtn: "Editar",
-    deleteDeviceBtn: "Eliminar",
-    cancelEditBtn: "Cancelar",
-    exportDataBtn: "Exportar Base de Datos",
-    importDataBtn: "Importar Base de Datos",
-    alertImportSuccess: "¡Base de datos importada! {num_devices} dispositivos cargados.",
-    alertImportError: "Error al importar la base de datos.",
-    confirmExportAndRevert: "¿Seguro que deseas exportar y restaurar la base de datos por defecto?",
-    alertExportAndRevertSuccess: "Base de datos exportada y restablecida.",
-
-    placeholder_deviceName: "Nombre del dispositivo",
-    placeholder_watt: "ej. 12.5",
-    placeholder_capacity: "ej. 98",
-    placeholder_pin: "ej. 10",
-    placeholder_dtap: "ej. 5",
-    placeholder_voltage: "11V-34V DC",
-    placeholder_port: "LEMO 8-pin",
-    placeholder_plates: '[{"type":"V-Mount","mount":"native"}]',
-    placeholder_media: '[{"type":"CFast 2.0"}]',
-    placeholder_lensmount: "ARRI PL",
-    placeholder_powerdist: "[{}]",
-    placeholder_videooutputs: "[{}]",
-    placeholder_fizconnector: "[{}]",
-    placeholder_timecode: "[{}]",
-
-    toggleDeviceManager: "Editar Datos de Dispositivos…",
-    hideDeviceManager: "Ocultar Datos de Dispositivos",
-
-    batteryTableLabel: "Batería",
-    runtimeLabel: "Autonomía Estimada (h)",
-    batteryLifeUnit: "h",
-
-    noBatterySupports: "Ninguna batería puede suministrar esta carga.",
-
-    alertSetupName: "Ingresa un nombre para la configuración.",
-    alertSetupSaved: 'Configuración "{name}" guardada.',
-    alertNoSetupSelected: "Selecciona una configuración para eliminar.",
-    confirmDeleteSetup: '¿Eliminar configuración "{name}"?',
-    confirmDeleteDevice: '¿Eliminar dispositivo "{name}"?',
-    alertDeviceExists: "Ya existe un dispositivo con ese nombre en esta categoría.",
-    alertDeviceAdded: 'Dispositivo "{name}" agregado a "{category}".',
-    alertDeviceUpdated: 'Dispositivo "{name}" actualizado en "{category}".',
-    alertDeviceFields: "Introduce valores válidos para capacidad, Pin A y D-Tap A.",
-    alertDeviceWatt: "Introduce un valor de vatios válido.",
-    alertDeviceName: "El nombre no puede estar vacío.",
-    alertInvalidCameraJSON: "JSON de cámara no válido",
-
-    setupActionsHeading: "Acciones de Configuración",
-    exportSetupsBtn: "Exportar Todas",
-    importSetupsBtn: "Importar Configuraciones",
-    generateOverviewBtn: "Generar Resumen",
-    alertNoSetupsToExport: "No hay configuraciones para exportar.",
-    alertImportSetupsSuccess: "{num_setups} configuraciones importadas.",
-    alertImportSetupsError: "Error: no se pudieron importar las configuraciones.",
-    alertSelectSetupForOverview: "Selecciona una configuración para generar un resumen.",
-    overviewTitle: "Resumen de Configuración",
-    exportAndRevertBtn: "Exportar y Restaurar por Defecto",
-    // NEW TEXTS FOR SETUP MANAGEMENT END HERE
-  },
-  fr: {
-    appTitle: "Planificateur de Consommation Caméra",
-    appHeading: "Application de Consommation Caméra",
-    tagline: "Planifiez votre configuration et calculez la consommation et l'autonomie.",
-
-    setupManageHeading: "Gérer la Configuration",
-    deviceSelectionHeading: "Sélection des Appareils",
-    resultsHeading: "Résultats",
-    deviceManagerHeading: "Gérer la Base de Données",
-    batteryComparisonHeading: "Comparaison des Batteries",
-
-    savedSetupsLabel: "Configurations Enregistrées:",
-    newSetupOption: "-- Nouvelle Configuration --",
-    setupNameLabel: "Nom de la Configuration:",
-    deleteSetupBtn: "Supprimer",
-    saveSetupBtn: "Enregistrer",
-
-    cameraLabel: "Caméra:",
-    monitorLabel: "Moniteur:",
-    videoLabel: "Vidéo Sans Fil:",
-    distanceLabel: "Capteur de Distance:",
-    batteryLabel: "Batterie V-Mount:",
-
-    fizLegend: "Systèmes FIZ",
-    fizMotorsLabel: "Moteurs FIZ:",
-    fizControllersLabel: "Contrôleurs FIZ:",
-
-    totalPowerLabel: "Consommation Totale:",
-    totalCurrent144Label: "Courant Total (à 14,4V):",
-    totalCurrent12Label: "Courant Total (à 12V):",
-    batteryLifeLabel: "Autonomie (estimée):",
-
-    warnPinExceeded: "ATTENTION: le courant ({current}A à ~12V) dépasse la limite principale ({max}A)!",
-    warnPinNear: "Remarque : le courant est proche de la limite ({current}A sur {max}A).",
-    warnDTapExceeded: "ATTENTION: le courant ({current}A) dépasse la sortie D-Tap ({max}A)!",
-    warnDTapNear: "Remarque : le courant utilise plus de 80% de la capacité D-Tap ({current}A sur {max}A).",
-
-    category_cameras: "Caméras",
-    category_monitors: "Moniteurs",
-    category_video: "Vidéo Sans Fil",
-    category_fiz_motors: "Moteurs FIZ",
-    category_fiz_controllers: "Contrôleurs FIZ",
-    category_fiz_distance: "Distance FIZ",
-    category_batteries: "Batteries V-Mount",
-
-    addDeviceHeading: "Ajouter un Nouvel Appareil",
-    categoryLabel: "Catégorie:",
-    deviceNameLabel: "Nom:",
-    consumptionLabel: "Consommation (W):",
-    capacityLabel: "Capacité (Wh):",
-    pinLabel: "Max Pin A:",
-    dtapLabel: "Max D-Tap A:",
-    cameraWattLabel: "Puissance (W):",
-    cameraVoltageLabel: "Plage de Tension:",
-    cameraPortTypeLabel: "Type de Port:",
-    cameraPlatesLabel: "Plates Batterie:",
-    cameraMediaLabel: "Supports d'Enregistrement:",
-    cameraLensMountLabel: "Monture d'Objectif:",
-    powerInputsHeading: "Entrées d'Alimentation",
-    powerDistributionHeading: "Distribution d'Alimentation",
-    videoOutputsHeading: "Sorties Vidéo",
-    fizConnectorHeading: "Connecteur FIZ",
-    mediaHeading: "Médias",
-    viewfinderHeading: "Viseur",
-    lensMountHeading: "Monture",
-    timecodeHeading: "Timecode",
-    powerDistributionLabel: "Sorties:",
-    videoOutputsLabel: "Sorties:",
-    fizConnectorLabel: "Connecteurs:",
-    viewfinderLabel: "Viseurs:",
-    timecodeLabel: "Timecode:",
-    addDeviceBtn: "Ajouter",
-    updateDeviceBtn: "Mettre à Jour",
-    editBtn: "Modifier",
-    deleteDeviceBtn: "Supprimer",
-    cancelEditBtn: "Annuler",
-    exportDataBtn: "Exporter la Base",
-    importDataBtn: "Importer la Base",
-    alertImportSuccess: "Base importée ! {num_devices} appareils chargés.",
-    alertImportError: "Échec de l'importation de la base.",
-    confirmExportAndRevert: "Exporter puis rétablir la base par défaut ?",
-    alertExportAndRevertSuccess: "Base exportée et réinitialisée.",
-
-    placeholder_deviceName: "Nom de l'appareil",
-    placeholder_watt: "ex. 12.5",
-    placeholder_capacity: "ex. 98",
-    placeholder_pin: "ex. 10",
-    placeholder_dtap: "ex. 5",
-    placeholder_voltage: "11V-34V DC",
-    placeholder_port: "LEMO 8-pin",
-    placeholder_plates: '[{"type":"V-Mount","mount":"native"}]',
-    placeholder_media: '[{"type":"CFast 2.0"}]',
-    placeholder_lensmount: "ARRI PL",
-    placeholder_powerdist: "[{}]",
-    placeholder_videooutputs: "[{}]",
-    placeholder_fizconnector: "[{}]",
-    placeholder_timecode: "[{}]",
-
-    toggleDeviceManager: "Éditer les Données…",
-    hideDeviceManager: "Masquer les Données",
-
-    batteryTableLabel: "Batterie",
-    runtimeLabel: "Autonomie Estimée (h)",
-    batteryLifeUnit: "h",
-
-    noBatterySupports: "Aucune batterie ne peut fournir cette charge.",
-
-    alertSetupName: "Veuillez saisir un nom pour la configuration.",
-    alertSetupSaved: 'Configuration "{name}" enregistrée.',
-    alertNoSetupSelected: "Sélectionnez une configuration à supprimer.",
-    confirmDeleteSetup: 'Supprimer la configuration "{name}" ?',
-    confirmDeleteDevice: "Supprimer l'appareil \"{name}\" ?",
-    alertDeviceExists: "Un appareil porte déjà ce nom dans cette catégorie.",
-    alertDeviceAdded: 'Appareil "{name}" ajouté à "{category}".',
-    alertDeviceUpdated: 'Appareil "{name}" mis à jour dans "{category}".',
-    alertDeviceFields: "Indiquez des valeurs valides pour capacité, Pin A et D-Tap A.",
-    alertDeviceWatt: "Veuillez entrer une valeur en watts valide.",
-    alertDeviceName: "Le nom de l'appareil ne peut être vide.",
-    alertInvalidCameraJSON: "JSON caméra invalide",
-
-    setupActionsHeading: "Actions de Configuration",
-    exportSetupsBtn: "Exporter Toutes",
-    importSetupsBtn: "Importer Configurations",
-    generateOverviewBtn: "Générer un Résumé",
-    alertNoSetupsToExport: "Aucune configuration à exporter.",
-    alertImportSetupsSuccess: "{num_setups} configurations importées.",
-    alertImportSetupsError: "Erreur : import impossible.",
-    alertSelectSetupForOverview: "Sélectionnez une configuration pour générer un résumé.",
-    overviewTitle: "Aperçu de Configuration",
-    exportAndRevertBtn: "Exporter et Réinitialiser",
-    // NEW TEXTS FOR SETUP MANAGEMENT END HERE
-  },
-  de: {
-    appTitle: "Kamera Stromverbrauchs Planer",
-    appHeading: "Kamera-Stromverbrauchs-App",
-    tagline: "Plane dein Kamera-Setup und berechne Stromverbrauch und Akkulaufzeit.",
-
-    setupManageHeading: "Setup verwalten",
-    deviceSelectionHeading: "Geräte-Auswahl",
-    resultsHeading: "Ergebnisse",
-    deviceManagerHeading: "Geräte-Datenbank verwalten",
-    batteryComparisonHeading: "Akkuvergleich",
-
-    savedSetupsLabel: "Gespeicherte Setups:",
-    newSetupOption: "-- Neues Setup --",
-    setupNameLabel: "Setup-Name:",
-    deleteSetupBtn: "Löschen",
-    saveSetupBtn: "Speichern",
-
-    cameraLabel: "Kamera:",
-    monitorLabel: "Monitor:",
-    videoLabel: "Videofunk:",
-    distanceLabel: "Distanzmesser:",
-    batteryLabel: "V-Mount Akku:",
-
-    fizLegend: "FIZ (Follow Focus) Systeme",
-    fizMotorsLabel: "FIZ Motoren:",
-    fizControllersLabel: "FIZ Controller:",
-
-    totalPowerLabel: "Gesamtverbrauch:",
-    totalCurrent144Label: "Gesamtstrom (bei 14,4V):",
-    totalCurrent12Label: "Gesamtstrom (bei 12V):",
-    batteryLifeLabel: "Akkulaufzeit (geschätzt):",
-
-    warnPinExceeded: "WARNUNG: Strombedarf ({current}A bei ~12V) übersteigt die Haupt-Pin-Ausgangsleistung ({max}A) des Akkus!",
-    warnPinNear: "Hinweis: Strombedarf kommt nahe an die Pin-Grenze ({current}A von {max}A).",
-    warnDTapExceeded: "WARNUNG: Strombedarf ({current}A) übersteigt die D-Tap-Leistung ({max}A) des Akkus!",
-    warnDTapNear: "Hinweis: Strombedarf nutzt über 80% der D-Tap Kapazität ({current}A von {max}A).",
-
-    category_cameras: "Kameras",
-    category_monitors: "Monitore",
-    category_video: "Videofunk",
-    category_fiz_motors: "FIZ Motoren",
-    category_fiz_controllers: "FIZ Controller",
-    category_fiz_distance: "FIZ Distanz",
-    category_batteries: "V-Mount Akkus",
-
-    addDeviceHeading: "Neues Gerät hinzufügen",
-    categoryLabel: "Kategorie:",
-    deviceNameLabel: "Name:",
-    consumptionLabel: "Verbrauch (W):",
-    capacityLabel: "Kapazität (Wh):",
-    pinLabel: "Max Pin A:",
-    dtapLabel: "Max D-Tap A:",
-    cameraWattLabel: "Leistungsaufnahme (W):",
-    cameraVoltageLabel: "Spannungsbereich:",
-    cameraPortTypeLabel: "Anschlussart:",
-    cameraPlatesLabel: "Akkuschächte:",
-    cameraMediaLabel: "Aufnahmemedien:",
-    cameraLensMountLabel: "Objektivanschluss:",
-    powerInputsHeading: "Stromversorgung",
-    powerDistributionHeading: "Stromverteilung",
-    videoOutputsHeading: "Videoausgänge",
-    fizConnectorHeading: "FIZ-Anschluss",
-    mediaHeading: "Medien",
-    viewfinderHeading: "Sucher",
-    lensMountHeading: "Objektivanschluss",
-    timecodeHeading: "Timecode",
-    powerDistributionLabel: "Ausgänge:",
-    videoOutputsLabel: "Ausgänge:",
-    fizConnectorLabel: "Anschlüsse:",
-    viewfinderLabel: "Sucher:",
-    timecodeLabel: "Timecode:",
-    addDeviceBtn: "Hinzufügen",
-    updateDeviceBtn: "Aktualisieren",
-    editBtn: "Bearbeiten",
-    deleteDeviceBtn: "Löschen",
-    cancelEditBtn: "Abbrechen",
-    exportDataBtn: "Datenbank exportieren",
-    importDataBtn: "Datenbank importieren",
-    alertImportSuccess: "Datenbank erfolgreich importiert! {num_devices} Geräte geladen.",
-    alertImportError: "Fehler beim Import der Datenbank. Ungültige Datei oder Datenformat.",
-    confirmExportAndRevert: "Möchten Sie die aktuelle Datenbank exportieren UND dann auf die Standarddatenbank zurücksetzen? Dies wird Ihre aktuell gespeicherten Daten mit den ursprünglichen Standardwerten überschreiben.",
-    alertExportAndRevertSuccess: "Datenbank exportiert und erfolgreich auf Standardwerte zurückgesetzt.",
-
-    placeholder_deviceName: "Gerätename",
-    placeholder_watt: "z.B. 12.5",
-    placeholder_capacity: "z.B. 98",
-    placeholder_pin: "z.B. 10",
-    placeholder_dtap: "z.B. 5",
-    placeholder_voltage: "11V-34V DC",
-    placeholder_port: "Bat LEMO 8-Pin",
-    placeholder_plates: '[{"type":"V-Mount","mount":"native"}]',
-    placeholder_media: '[{"type":"CFast 2.0"}]',
-    placeholder_lensmount: "ARRI PL",
-    placeholder_powerdist: "[{}]",
-    placeholder_videooutputs: "[{}]",
-    placeholder_fizconnector: "[{}]",
-    placeholder_timecode: "[{}]",
-
-    toggleDeviceManager: "Gerätedaten bearbeiten…",
-    hideDeviceManager: "Gerätedaten ausblenden",
-
-    batteryTableLabel: "Akku",
-    runtimeLabel: "Geschätzte Laufzeit (h)",
-    batteryLifeUnit: "Std.",
-
-    noBatterySupports: "Kein Akku kann diese Last liefern.",
-
-    alertSetupName: "Bitte einen Namen für das Setup eingeben.",
-    alertSetupSaved: 'Setup "{name}" gespeichert.',
-    alertNoSetupSelected: "Wählen Sie ein gespeichertes Setup zum Löschen aus.",
-    confirmDeleteSetup: 'Setup "{name}" wirklich löschen?',
-    confirmDeleteDevice: 'Gerät "{name}" wirklich löschen?',
-    alertDeviceExists: "Ein Gerät mit diesem Namen existiert in dieser Kategorie bereits.",
-    alertDeviceAdded: 'Gerät "{name}" in Kategorie "{category}" hinzugefügt.',
-    alertDeviceUpdated: 'Gerät "{name}" in Kategorie "{category}" aktualisiert.',
-    alertDeviceFields: "Bitte geben Sie gültige Werte für Kapazität, Pin A und D-Tap A ein.",
-    alertDeviceWatt: "Bitte geben Sie einen gültigen Watt-Wert ein.",
-    alertDeviceName: "Der Gerätename darf nicht leer sein.",
-    alertInvalidCameraJSON: "Ungültiges JSON für Kameradetails",
-    // NEW TEXTS FOR SETUP MANAGEMENT START HERE
-    setupActionsHeading: "Setup-Aktionen",
-    exportSetupsBtn: "Alle Setups exportieren",
-    importSetupsBtn: "Setups importieren",
-    generateOverviewBtn: "Übersicht erstellen",
-    alertNoSetupsToExport: "Es gibt keine gespeicherten Setups zum Exportieren.",
-    alertImportSetupsSuccess: "{num_setups} Setups erfolgreich importiert.",
-    alertImportSetupsError: "Fehler: Setups konnten nicht importiert werden. Die Datei ist möglicherweise ungültig oder beschädigt.",
-    alertSelectSetupForOverview: "Bitte wählen Sie ein gespeichertes Setup, um eine Übersicht zu erstellen.",
-    overviewTitle: "Setup-Übersicht",
-    exportAndRevertBtn: "Exportieren und auf Standarddatenbank zurücksetzen",
-    // NEW TEXTS FOR SETUP MANAGEMENT END HERE
+// Load translations when not already present (mainly for tests)
+if (typeof texts === 'undefined') {
+  try {
+    const translations = require('./translations.js');
+    window.texts = translations.texts;
+    window.categoryNames = translations.categoryNames;
+  } catch (e) {
+    console.warn('Failed to load translations', e);
   }
-};
+}
 
-// Mapping for category values to display names (singular for form dropdown)
-const categoryNames = {
-  en: {
-    "cameras": "Camera",
-    "monitors": "Monitor",
-    "video": "Wireless Video",
-    "fiz.motors": "FIZ Motor",
-    "fiz.controllers": "FIZ Controller",
-    "fiz.distance": "FIZ Distance",
-    "batteries": "V-Mount Battery"
-  },
-  de: {
-    "cameras": "Kamera",
-    "monitors": "Monitor",
-    "video": "Videofunk",
-    "fiz.motors": "FIZ Motoren",
-    "fiz.controllers": "FIZ Controller",
-    "fiz.distance": "FIZ Distanz",
-    "batteries": "V-Mount Akku"
-  },
-  es: {
-    "cameras": "Cámara",
-    "monitors": "Monitor",
-    "video": "Video inalámbrico",
-    "fiz.motors": "Motor FIZ",
-    "fiz.controllers": "Control FIZ",
-    "fiz.distance": "Distancia FIZ",
-    "batteries": "Batería V-Mount"
-  },
-  fr: {
-    "cameras": "Caméra",
-    "monitors": "Moniteur",
-    "video": "Vidéo sans fil",
-    "fiz.motors": "Moteur FIZ",
-    "fiz.controllers": "Contrôleur FIZ",
-    "fiz.distance": "Distance FIZ",
-    "batteries": "Batterie V-Mount"
-  }
-};
 
 // Determine initial language (default English)
 let currentLang = "en";
@@ -1085,6 +524,7 @@ function updateFizConnectorOptions() {
   });
 }
 
+// Wrap a form field with a div containing a data-label attribute for styling.
 function createFieldWithLabel(el, label) {
   const wrapper = document.createElement('div');
   wrapper.className = 'field-with-label';
@@ -1093,6 +533,7 @@ function createFieldWithLabel(el, label) {
   return wrapper;
 }
 
+// Helper used by select-row builders to insert an empty option.
 function addEmptyOption(select) {
   const opt = document.createElement("option");
   opt.value = "";
@@ -1100,6 +541,7 @@ function addEmptyOption(select) {
   select.appendChild(opt);
 }
 
+// Build a single row of the video output editor UI.
 function createVideoOutputRow(value = '') {
   const row = document.createElement('div');
   row.className = 'form-row';
@@ -1151,6 +593,7 @@ function clearVideoOutputs() {
   setVideoOutputs([]);
 }
 
+// Build a row for editing a FIZ connector entry.
 function createFizConnectorRow(value = '') {
   const row = document.createElement('div');
   row.className = 'form-row';
@@ -1230,6 +673,7 @@ function updateRecordingMediaOptions() {
   });
 }
 
+// Build a row allowing the user to specify recording media details.
 function createRecordingMediaRow(type = '', notes = '') {
   const row = document.createElement('div');
   row.className = 'form-row';
@@ -1362,6 +806,7 @@ function updatePlateTypeOptions() {
   });
 }
 
+// Build a battery plate row with type, mount and optional notes fields.
 function createBatteryPlateRow(type = '', mount = 'native', notes = '') {
   const row = document.createElement('div');
   row.className = 'form-row';
@@ -1470,6 +915,7 @@ function getAllViewfinderConnectors() {
 let viewfinderTypeOptions = getAllViewfinderTypes();
 let viewfinderConnectorOptions = getAllViewfinderConnectors();
 
+// Build a viewfinder configuration row used in the camera editor.
 function createViewfinderRow(type = '', resolution = '', connector = '', notes = '') {
   const row = document.createElement('div');
   row.className = 'form-row';
@@ -1599,6 +1045,7 @@ function updateMountTypeOptions() {
   });
 }
 
+// Build a lens mount row with type and mount selection fields.
 function createLensMountRow(type = '', mount = 'native') {
   const row = document.createElement('div');
   row.className = 'form-row';
@@ -1759,6 +1206,7 @@ function updatePowerDistTypeOptions() {
   });
 }
 
+// Build a power distribution output row for the editor UI.
 function createPowerDistRow(type = '', voltage = '', current = '', wattage = '', notes = '') {
   const row = document.createElement('div');
   row.className = 'form-row';
@@ -1907,6 +1355,7 @@ function updateTimecodeTypeOptions() {
   });
 }
 
+// Build a timecode connector row used for editing camera properties.
 function createTimecodeRow(type = '', notes = '') {
   const row = document.createElement('div');
   row.className = 'form-row';
